app/requirements.txt
flask
passlib
flask-wtf
gunicorn
psycopg2-binary